merge

+++ src/Visualization/TextEditor.java
@@ -3,7 +3,7 @@ package Visualization;
 import javafx.scene.control.TextArea;
 
 public class TextEditor extends TextArea{
-    public TextEditor(int width, int height){
+    public TextEditor(double width, double height){
         setMaxWidth(width/2 - 30);
         setPrefHeight(height * 0.75);
     }
